firstCatch works

+++ utils/monsterUtils.js
@@ -1,18 +1,19 @@
 import { getFirestore, collection, query, onSnapshot, where, getDocs } from 'firebase/firestore';
 import { getStorage, ref, getDownloadURL } from 'firebase/storage';
 import AsyncStorage from '@react-native-async-storage/async-storage';
+import AchievementAlert from '../modals/AchievementAlert';
 
 export const findMonster = () => {
   const randomNumber = Math.random() * 100;
   let foundMonsterId;
 
-  if (randomNumber < 2) {
+  if (randomNumber < 10) {
     // Epic monster
     foundMonsterId = Math.floor(Math.random() * 5) + 46; // ID:t 46-50
-  } else if (randomNumber < 10) {
+  } else if (randomNumber < 40) {
     // Rare monster
     foundMonsterId = Math.floor(Math.random() * 15) + 31; // ID:t 31-45
-  } else if (randomNumber < 40) {
+  } else if (randomNumber < 100) {
     // Common monster
     foundMonsterId = Math.floor(Math.random() * 30) + 1; // ID:t 1-30
   }
@@ -26,20 +27,54 @@ export const saveMonsterToAsyncStorage = async (monster, userId) => {
     const existingMonsters = await AsyncStorage.getItem(`monsters_${userId}`);
     // Parse existing monsters or initialize an empty array
     const parsedExistingMonsters = existingMonsters ? JSON.parse(existingMonsters) : [];
-    // Add the new monster to the array
-    parsedExistingMonsters.push(monster);
-    // Save the updated monsters array to AsyncStorage
-    await AsyncStorage.setItem(`monsters_${userId}`, JSON.stringify(parsedExistingMonsters));
-  //  await AsyncStorage.setItem(`caughtMonsters_${userId}`, parsedExistingMonsters.length.toString());
-  //  console.log(`MONSTERUTILS: ${parsedExistingMonsters.length.toString()} monsters caught for user ID: ${userId}`);
-  //  console.log(`MONSTERUTILS 1: Saved monster to AsyncStorage: ${monster.name} for user ID: ${userId}`);
-
+    
+    // Extract existing uniqueMonsterIds or initialize an empty set
+    const existingUniqueMonsterIds = new Set(await AsyncStorage.getItem(`uniqueMonsterIds_${userId}`) || []);
+ 
+    const monsterId = monster.id;
+ 
+    // Check if the monsterId is already in the set
+    if (!existingUniqueMonsterIds.has(monsterId)) {
+      // Add the new monster to the array
+      parsedExistingMonsters.push(monster);
+      // Add the monsterId to the set
+      existingUniqueMonsterIds.add(monsterId);
+ 
+      // Save the updated monsters array and uniqueMonsterIds set to AsyncStorage
+      await AsyncStorage.setItem(`monsters_${userId}`, JSON.stringify(parsedExistingMonsters));
+      await AsyncStorage.setItem(`uniqueMonsterIds_${userId}`, JSON.stringify(Array.from(existingUniqueMonsterIds)));
+    }
+ 
+    // Update the user's progress towards the achievements
+    const userProgress = await AsyncStorage.getItem(`userProgress_${userId}`);
+    const parsedUserProgress = userProgress ? JSON.parse(userProgress) : {};
+ 
+    if (parsedUserProgress[monsterId]) {
+      parsedUserProgress[monsterId]++;
+    } else {
+      parsedUserProgress[monsterId] = 1;
+    }
+ 
+    // Log the updated user progress
+    console.log(`Updated user progress for monster ${monsterId}:`, parsedUserProgress[monsterId]);
+ 
+    await AsyncStorage.setItem(`userProgress_${userId}`, JSON.stringify(parsedUserProgress));
+ 
+    // Check if the user has achieved any of the achievements
+    const newAchievements = await checkAchievements(userId);
+ 
     // Trigger the callback to update the component
+    if (newAchievements.length > 0) {
+      // If there are new achievements, display an alert
+      const achievementMessage = `New Achievements: ${newAchievements.join(', ')}`;
+      AchievementAlert.show(achievementMessage);
+    }
+ 
   } catch (error) {
     console.error('Error saving monster to AsyncStorage:', error);
     throw error;
   }
-};
+ };
 
 /*
 export const saveImageToAsyncStorage = async (imageUrl, userId) => {
@@ -168,18 +203,203 @@ export const fetchMonsterImageURL = async (monsterId, userId) => {
   }
 };
 
-export const fetchAchievements = async (setAchievements) => {
+export const fetchAchievements = async () => {
   try {
     const db = getFirestore();
     const querySnapshot = await getDocs(collection(db, 'achievements'));
     const achievements = [];
+
     querySnapshot.forEach((doc) => {
-      achievements.push(doc.data());
+      const achievementData = doc.data();
+      const achievement = {
+        id: doc.id,
+        name: achievementData.name,
+        description: achievementData.description,
+        trigger: achievementData.trigger,
+      };
+
+      achievements.push(achievement);
     });
-    setAchievements(achievements);
-    console.log(achievements);
+
+    return achievements;
   } catch (error) {
     console.error('Error reading achievements from Firestore: ', error);
     throw error;
   }
- };
+};
+
+export const checkAchievements = async (userId) => {
+  try {
+    // Retrieve user progress from AsyncStorage
+    const userProgress = await AsyncStorage.getItem(`userProgress_${userId}`);
+    const parsedUserProgress = userProgress ? JSON.parse(userProgress) : {};
+
+    // Retrieve achievements from Firestore
+    const achievements = await fetchAchievements();
+
+    // Check each achievement
+    achievements.forEach(async (achievement) => {
+      const triggerType = achievement.trigger.type;
+      const triggerCount = achievement.trigger.count;
+      const triggerMonsterRange = achievement.trigger.monsterRange;
+      const triggerSpecificMonsters = achievement.trigger.specificMonsters;
+
+      // Implement different checks based on the trigger type
+      switch (triggerType) {
+        case 'collectCount':
+          const collectedCount = calculateCollectedCount(parsedUserProgress, triggerMonsterRange);
+          if (collectedCount >= triggerCount) {
+            // Achievement unlocked
+            console.log(`Achievement unlocked: ${achievement.name}`);
+            // Add logic to notify the user or update UI as needed
+          }
+          break;
+
+        case 'collectSpecific':
+          const collectedSpecificCount = calculateCollectedSpecific(parsedUserProgress, triggerSpecificMonsters);
+          if (collectedSpecificCount >= triggerCount) {
+            // Achievement unlocked
+            console.log(`Achievement unlocked: ${achievement.name}`);
+            // Add logic to notify the user or update UI as needed
+          }
+          break;
+
+        case 'collectAll':
+          const collectedAllCount = calculateCollectedAll(parsedUserProgress, triggerMonsterRange);
+          console.log(`Collected all count: ${collectedAllCount}`);
+          if (collectedAllCount >= triggerCount) {
+            // Achievement unlocked
+            console.log(`Achievement unlocked: ${achievement.name}`);
+            // Add logic to notify the user or update UI as needed
+          }
+          break;
+
+        case 'firstCatch':
+          const hasCaughtFirstMonster = calculateFirstCatch(parsedUserProgress, triggerMonsterRange);
+          if (hasCaughtFirstMonster) {
+            // Achievement unlocked
+            console.log(`Achievement unlocked: ${achievement.name}`);
+            // Add logic to notify the user or update UI as needed
+          }
+          break;
+
+        // Add more cases for other trigger types if needed
+
+        default:
+          break;
+      }
+    });
+    return [];
+  } catch (error) {
+    console.error('Error checking achievements:', error);
+    //return []; // Return an empty array in case of an error
+  }
+};
+
+const calculateCollectedCount = async (userId, monsterRange) => {
+  let collectedCount = 0;
+  
+  // Retrieve uniqueMonsterIds set from AsyncStorage
+  const uniqueMonsterIds = new Set(await AsyncStorage.getItem(`uniqueMonsterIds_${userId}`) || []);
+  //console.log(uniqueMonsterIds)
+  for (let i = monsterRange[0]; i <= monsterRange[1]; i++) {
+    const monsterId = i.toString();
+    if (uniqueMonsterIds.has(monsterId)) {
+      collectedCount += userProgress[monsterId] || 0;
+    }
+  }
+
+  return collectedCount;
+};
+
+// Updated calculateCollectedSpecific function to use uniqueMonsterIds set
+const calculateCollectedSpecific = async (userId, specificMonsters) => {
+  let collectedSpecificCount = 0;
+
+  // Retrieve uniqueMonsterIds set from AsyncStorage
+  const uniqueMonsterIds = new Set(await AsyncStorage.getItem(`uniqueMonsterIds_${userId}`) || []);
+
+  specificMonsters.forEach((monsterId) => {
+    const stringMonsterId = monsterId.toString();
+    if (uniqueMonsterIds.has(stringMonsterId)) {
+      collectedSpecificCount += userProgress[stringMonsterId] || 0;
+    }
+  });
+
+  return collectedSpecificCount;
+};
+
+const calculateCollectedAll = async (userId, monsterRange) => {
+  // Retrieve uniqueMonsterIds set from AsyncStorage
+  const uniqueMonsterIds = new Set(await AsyncStorage.getItem(`uniqueMonsterIds_${userId}`) || []);
+
+  for (let i = monsterRange[0]; i <= monsterRange[1]; i++) {
+    const monsterId = i.toString();
+    console.log(`Monster ID: ${monsterId}`);
+    if (!uniqueMonsterIds.has(monsterId) || userProgress[monsterId] < 1) {
+      return 0; // Not all monsters in the range are collected
+    }
+  }
+
+  return 1; // All monsters in the range are collected
+};
+
+// Updated calculateFirstCatch function to use uniqueMonsterIds set
+const calculateFirstCatch = (userProgress, monsterRange) => {
+  for (let i = monsterRange[0]; i <= monsterRange[1]; i++) {
+    const monsterId = i.toString();
+    if (userProgress[monsterId] && userProgress[monsterId] > 0) {
+      return true; // User has caught at least one monster in the range
+    }
+  }
+
+  return false; // User has not caught any monster in the range
+};
+/*export const checkAchievements = async (userId) => {
+  try {
+   // Fetch the user's progress from AsyncStorage
+   const userProgressString = await AsyncStorage.getItem(`userProgress_${userId}`);
+   const userProgress = userProgressString ? JSON.parse(userProgressString) : {};
+ 
+   // Fetch the achievements from Firestore
+   const achievements = await fetchAchievements();
+ 
+   // Check if the user has achieved any of the achievements
+   achievements.forEach(achievement => {
+     if (achievement.trigger.type === "firstCatch") {
+       // Check if the user has caught any monster within the specified range
+       const hasCaughtAnyMonster = Object.keys(userProgress)
+         .some(monsterId => achievement.trigger.monsterRange.includes(Number(monsterId)));
+ 
+       if (hasCaughtAnyMonster) {
+         console.log(`User has achieved: ${achievement.name}`);
+       }
+     } else if (achievement.trigger.type === "collectCount") {
+       // Existing code...
+     }
+     else if (achievement.trigger.type === "collectSpecific") {
+      // Check if the user has caught all specific monsters
+      const hasCaughtAllSpecificMonsters = achievement.trigger.specificMonsters.every(
+        monsterId => userProgress[monsterId]
+      );
+     
+      if (hasCaughtAllSpecificMonsters) {
+        console.log(`User has achieved: ${achievement.name}`);
+      }
+     }
+     else if (achievement.trigger.type === "collectAll") {
+      // Check if the user has caught all monsters within the range
+      const hasCaughtAllMonstersInRange = achievement.trigger.monsterRange.every(
+        monsterId => userProgress[monsterId] === true
+       );
+     
+      if (hasCaughtAllMonstersInRange) {
+        console.log(`User has achieved: ${achievement.name}`);
+      }
+     }
+   });
+  } catch (error) {
+   console.error('Error checking achievements:', error);
+   throw error;
+  }
+ };*/
